+++ praw/models/reddit/redditor.py
@@ -40,6 +40,7 @@ class Redditor(RedditBase, MessageableMixin, RedditorListingMixin):
 
     def _friend(self, method, data):
         url = API_PATH['friend_v1'].format(user=self)
+        # PRAW5 REMOVE (return statement)
         return self._reddit.request(method, url, data=dumps(data))
 
     def friend(self, note=None):
@@ -47,11 +48,11 @@ class Redditor(RedditBase, MessageableMixin, RedditorListingMixin):
 
         :param note: A note to save along with the relationship. Requires
             reddit Gold (default: None).
-        :returns: The json response from the server.
 
         Calling this method subsequent times will update the note.
 
         """
+        # PRAW5 REMOVE (return statement)
         return self._friend('PUT', data={'note': note} if note else {})
 
     def friend_info(self):
@@ -82,8 +83,6 @@ class Redditor(RedditBase, MessageableMixin, RedditorListingMixin):
     def unblock(self):
         """Unblock the Redditor.
 
-        :returns: The json response from the server.
-
         Blocking must be done from a Message, Comment Reply or Submission
         Reply.
 
@@ -91,6 +90,7 @@ class Redditor(RedditBase, MessageableMixin, RedditorListingMixin):
         data = {'container': self._reddit.user.me().fullname,
                 'name': str(self), 'type': 'enemy'}
         url = API_PATH['unfriend'].format(subreddit='all')
+        # PRAW5 REMOVE (return statement)
         return self._reddit.post(url, data=data)
 
     def unfriend(self):