+++ CHANGES.rst
@@ -6,6 +6,9 @@ Unreleased
 **Added**
 
 * :meth:`.Subreddit.rules` to get the rules of a subreddit.
+* :class:`.LiveContributorRelationship`, which can be obtained through
+  :attr:`.LiveThread.contributor`, to interact with live threads'
+  contributors.
 
 **Fixed**