+++ CHANGES.rst
@@ -10,6 +10,9 @@ Unreleased
   :attr:`.LiveThread.contributor`, to interact with live threads'
   contributors.
 * :meth:`~.remove_invite` to remove a moderator invite.
+* :meth:`.LiveContributorRelationship.invite` to send a contributor invitation.
+* :meth:`.LiveContributorRelationship.remove_invite` to remove the contributor
+  invitation.
 
 **Fixed**